router: protect replicasets from modification
The function `vshard.router.routeall` is not marked as internal API, but
it returns `router.replicasets` table, which is used by the router itself.
And also `vshard.router.route` returns a replicaset object that is used
internally. If user modifies the returned from `routeall` or `route`
table, then router is broken and cannot continue to work properly after
that.

Therefore, it was decided to return protected tables from these methods.
Now `routeall` returns a table with replicasets that doesn't support
adding or deleting. And also all replicaset objects that are returned to
the user from `routeall` or `route` only support `call`, `callrw`,
`callro`, `callbro`, `callre`, `callbre` calls and do not provide direct
access to their fields.

Closes #502

NO_DOC=<undocumented part>